add a call on the send page as well
