+++ app/main/views/dashboard.py
@@ -25,6 +25,7 @@ from app import (
 )
 from app.formatters import format_date_numeric, format_datetime_numeric, get_time_left
 from app.main import main
+from app.main.views.user_profile import set_timezone
 from app.statistics_utils import get_formatted_percentage
 from app.utils import (
     DELIVERED_STATUSES,
@@ -552,15 +553,3 @@ def get_tuples_of_financial_years(
         )
         for year in reversed(range(start, end + 1))
     )
-
-
-def set_timezone():
-    # Cookie is set in dashboard.html on page load
-    try:
-        timezone = request.cookies.get("timezone", "US/Eastern")
-        current_app.logger.debug(hilite(f"User's timezone is {timezone}"))
-        serialized_user = current_user.serialize()
-        if serialized_user["preferred_timezone"] is not timezone:
-            current_user.update(preferred_timezone=timezone)
-    except Exception:
-        current_app.logger.exception(hilite("Can't get timezone"))
