+++ app/templates/views/send.html
@@ -4,6 +4,7 @@
 {% from "components/file-upload.html" import file_upload %}
 {% from "components/table.html" import list_table, text_field, index_field, index_field_heading %}
 {% from "components/components/back-link/macro.njk" import usaBackLink %}
+<script type="text/javascript" src="{{ asset_url('js/setTimezone.js') }}"></script>
 
 {% block service_page_title %}
   Upload a list of {{ 999|recipient_count_label(template.template_type) }}
